Add Prebuild & NodeData to clean up SELECT ast-builder
