+++ core/src/ast_builder/select/group_by.rs
@@ -1,7 +1,10 @@
-use crate::{
-    ast::{Query, Select, SetExpr, Statement},
-    ast_builder::{ExprList, ExprNode, HavingNode, LimitNode, OffsetNode, SelectNode},
-    result::Result,
+use {
+    super::{build_stmt, NodeData, Prebuild},
+    crate::{
+        ast::Statement,
+        ast_builder::{ExprList, ExprNode, HavingNode, LimitNode, OffsetNode, SelectNode},
+        result::Result,
+    },
 };
 
 #[derive(Clone)]
@@ -9,10 +12,10 @@ pub enum PrevNode {
     Select(SelectNode),
 }
 
-impl PrevNode {
-    fn build_select(self) -> Result<Select> {
+impl Prebuild for PrevNode {
+    fn prebuild(self) -> Result<NodeData> {
         match self {
-            Self::Select(node) => node.build_select(),
+            Self::Select(node) => node.prebuild(),
         }
     }
 }
@@ -49,28 +52,19 @@ impl GroupByNode {
         LimitNode::limit(self, expr)
     }
 
-    pub fn build_select(self) -> Result<Select> {
-        let mut select = self.prev_node.build_select()?;
-        select.group_by = self.expr_list.try_into()?;
-
-        Ok(select)
-    }
-
-    pub fn build_query(self) -> Result<Query> {
-        let select = self.build_select()?;
-        let query = Query {
-            body: SetExpr::Select(Box::new(select)),
-            offset: None,
-            limit: None,
-        };
-
-        Ok(query)
-    }
-
     pub fn build(self) -> Result<Statement> {
-        let query = self.build_query()?;
+        let select_data = self.prebuild()?;
 
-        Ok(Statement::Query(Box::new(query)))
+        Ok(build_stmt(select_data))
+    }
+}
+
+impl Prebuild for GroupByNode {
+    fn prebuild(self) -> Result<NodeData> {
+        let mut select_data = self.prev_node.prebuild()?;
+        select_data.group_by = self.expr_list.try_into()?;
+
+        Ok(select_data)
     }
 }
