+++ core/src/ast_builder/select/mod.rs
@@ -4,144 +4,58 @@ mod limit;
 mod limit_offset;
 mod offset;
 mod offset_limit;
+mod root;
 
 pub use {
     group_by::GroupByNode, having::HavingNode, limit::LimitNode, limit_offset::LimitOffsetNode,
-    offset::OffsetNode, offset_limit::OffsetLimitNode,
+    offset::OffsetNode, offset_limit::OffsetLimitNode, root::SelectNode,
 };
 
-use {
-    super::{ExprList, ExprNode},
-    crate::{
-        ast::{
-            Expr, ObjectName, Query, Select, SelectItem, SetExpr, Statement, TableFactor,
-            TableWithJoins,
-        },
-        result::Result,
-    },
+use crate::{
+    ast::{Expr, Query, Select, SelectItem, SetExpr, Statement, TableWithJoins},
+    result::Result,
 };
 
-#[derive(Clone)]
-pub struct SelectNode {
-    table_name: String,
-    columns: Vec<String>,
-    filter_expr: Option<ExprNode>,
+struct NodeData {
+    pub projection: Vec<SelectItem>,
+    pub from: TableWithJoins,
+    /// WHERE
+    pub selection: Option<Expr>,
+    pub group_by: Vec<Expr>,
+    pub having: Option<Expr>,
+    pub limit: Option<Expr>,
+    pub offset: Option<Expr>,
 }
 
-impl SelectNode {
-    pub fn new(table_name: String) -> Self {
-        Self {
-            table_name,
-            columns: vec![],
-            filter_expr: None,
-        }
-    }
-
-    pub fn col(mut self, column: &str) -> Self {
-        self.columns.push(column.to_owned());
-
-        self
-    }
-
-    pub fn filter<T: Into<ExprNode>>(mut self, expr: T) -> Self {
-        self.filter_expr = Some(expr.into());
-
-        self
-    }
-
-    pub fn group_by<T: Into<ExprList>>(self, expr_list: T) -> GroupByNode {
-        GroupByNode::group_by(self, expr_list)
-    }
-
-    pub fn offset<T: Into<ExprNode>>(self, expr: T) -> OffsetNode {
-        OffsetNode::offset(self, expr)
-    }
-
-    pub fn limit<T: Into<ExprNode>>(self, expr: T) -> LimitNode {
-        LimitNode::limit(self, expr)
-    }
-
-    pub fn build_select(self) -> Result<Select> {
-        let relation = TableFactor::Table {
-            name: ObjectName(vec![self.table_name]),
-            alias: None,
-            index: None,
-        };
-
-        let from = TableWithJoins {
-            relation,
-            joins: vec![],
-        };
-
-        let selection = self.filter_expr.map(Expr::try_from).transpose()?;
-
-        let select = Select {
-            projection: vec![SelectItem::Wildcard],
-            from,
-            selection,
-            group_by: vec![],
-            having: None,
-            order_by: vec![],
-        };
-
-        Ok(select)
-    }
-
-    pub fn build_query(self) -> Result<Query> {
-        let select = self.build_select()?;
-        let query = Query {
-            body: SetExpr::Select(Box::new(select)),
-            offset: None,
-            limit: None,
-        };
-
-        Ok(query)
-    }
-
-    pub fn build(self) -> Result<Statement> {
-        let query = self.build_query()?;
-
-        Ok(Statement::Query(Box::new(query)))
-    }
+trait Prebuild {
+    fn prebuild(self) -> Result<NodeData>;
 }
 
-#[cfg(test)]
-mod tests {
-    use crate::{
-        ast::Statement, ast_builder::Builder, parse_sql::parse, result::Result,
-        translate::translate,
+fn build_stmt(select_data: NodeData) -> Statement {
+    let NodeData {
+        projection,
+        from,
+        selection,
+        group_by,
+        having,
+        offset,
+        limit,
+    } = select_data;
+
+    let select = Select {
+        projection,
+        from,
+        selection,
+        group_by,
+        having,
+        order_by: vec![],
     };
 
-    fn stmt(sql: &str) -> Result<Statement> {
-        let parsed = &parse(sql).unwrap()[0];
+    let query = Query {
+        body: SetExpr::Select(Box::new(select)),
+        offset,
+        limit,
+    };
 
-        translate(parsed)
-    }
-
-    #[test]
-    fn select() {
-        let actual = Builder::table("App").select().build();
-        let expected = stmt("SELECT * FROM App");
-        assert_eq!(actual, expected);
-
-        let actual = Builder::table("Bar").select().filter("id IS NULL").build();
-        let expected = stmt("SELECT * FROM Bar WHERE id IS NULL");
-        assert_eq!(actual, expected);
-
-        let actual = Builder::table("Foo")
-            .select()
-            .filter("id > name")
-            .offset(3)
-            .limit(100)
-            .build();
-        let expected = stmt(
-            "
-            SELECT * FROM Foo
-            WHERE id > name
-            OFFSET 3
-            LIMIT 100
-        ",
-        );
-        assert_eq!(actual, expected);
-    }
+    Statement::Query(Box::new(query))
 }
