+++ core/src/ast_builder/select/offset_limit.rs
@@ -1,7 +1,10 @@
-use crate::{
-    ast::{Query, Statement},
-    ast_builder::{ExprNode, GroupByNode, HavingNode, OffsetNode, SelectNode},
-    result::Result,
+use {
+    super::{build_stmt, NodeData, Prebuild},
+    crate::{
+        ast::Statement,
+        ast_builder::{ExprNode, GroupByNode, HavingNode, OffsetNode, SelectNode},
+        result::Result,
+    },
 };
 
 #[derive(Clone)]
@@ -12,13 +15,13 @@ pub enum PrevNode {
     Offset(OffsetNode),
 }
 
-impl PrevNode {
-    fn build_query(self) -> Result<Query> {
+impl Prebuild for PrevNode {
+    fn prebuild(self) -> Result<NodeData> {
         match self {
-            Self::Select(node) => node.build_query(),
-            Self::GroupBy(node) => node.build_query(),
-            Self::Having(node) => node.build_query(),
-            Self::Offset(node) => node.build_query(),
+            Self::Select(node) => node.prebuild(),
+            Self::GroupBy(node) => node.prebuild(),
+            Self::Having(node) => node.prebuild(),
+            Self::Offset(node) => node.prebuild(),
         }
     }
 }
@@ -62,9 +65,17 @@ impl OffsetLimitNode {
     }
 
     pub fn build(self) -> Result<Statement> {
-        let mut query = self.prev_node.build_query()?;
-        query.limit = Some(self.expr.try_into()?);
+        let select_data = self.prebuild()?;
 
-        Ok(Statement::Query(Box::new(query)))
+        Ok(build_stmt(select_data))
+    }
+}
+
+impl Prebuild for OffsetLimitNode {
+    fn prebuild(self) -> Result<NodeData> {
+        let mut select_data = self.prev_node.prebuild()?;
+        select_data.limit = Some(self.expr.try_into()?);
+
+        Ok(select_data)
     }
 }
